Json操作クラスを追加

+++ 2025年前期/GameSources/Map.h
@@ -36,11 +36,11 @@ namespace basecross{
 		void Load();
 
 		Vec3 GetMapCenter() {
-			Vec2 size = Vec2(m_Map[0].size() - 1, m_Map.size() - 1);
+			Vec2 size = Vec2(static_cast<float>(m_Map[0].size() - 1), static_cast<float>(m_Map.size() - 1));
 			return Vec3(size.x / 2.0f, m_CenterY, size.x / 2.0f);
 		}
 		Vec3 GetMapSize() {
-			return Vec3(m_Map[0].size(), m_MapHeight, m_Map.size());
+			return Vec3(static_cast<float>(m_Map[0].size()), m_MapHeight, static_cast<float>(m_Map.size()));
 		}
 
 		vector<Col4> GetColorTable() {
